Megaplan: parental settings split, TMDB people enrich, structured extractor errors
- Move parental PIN/lock UI into ParentalSettingsController (mobile+TV)
- Enrich AniWorld/SerienStream people via TmdbUtils.enrichPersonByName
- ExtractionFailedException after primary+fallback extractor failures
- CHANGELOG notes for provider/extractor fixes; restore gradlew executable bit
